services/orchestator/requirements.txt
llama-index-embeddings-text-embeddings-inference==0.1.2
llama-index-llms-ollama==0.1.3
llama-index-vector-stores-weaviate==1.0.0
llama-index-core==0.10.36
llama-index==0.10.36
weaviate-client==4.6.0
fastapi
uvicorn
ollama